+++ core/apis_test.go
@@ -30,7 +30,7 @@ import (
 
 func TestSerGetVersion(t *testing.T) {
 	storage := genmocks.Storage{}
-	coreAPIs := core.NewCoreAPIs("local", "1.1.1", "build", &storage, nil, nil, nil)
+	coreAPIs := core.NewCoreAPIs("local", "1.1.1", "build", &storage, nil, nil, nil, "", "", nil)
 	got := coreAPIs.GetVersion()
 	want := "1.1.1"
 
@@ -39,7 +39,7 @@ func TestSerGetVersion(t *testing.T) {
 
 func TestSerGetAuthTest(t *testing.T) {
 	storage := genmocks.Storage{}
-	coreAPIs := core.NewCoreAPIs("local", "1.1.1", "build", &storage, nil, nil, nil)
+	coreAPIs := core.NewCoreAPIs("local", "1.1.1", "build", &storage, nil, nil, nil, "", "", nil)
 
 	l := logs.NewLogger("test", nil).NewLog("1", logs.RequestContext{})
 	got := coreAPIs.Services.SerGetAuthTest(l)
@@ -50,7 +50,7 @@ func TestSerGetAuthTest(t *testing.T) {
 
 func TestSerGetCommonTest(t *testing.T) {
 	storage := genmocks.Storage{}
-	coreAPIs := core.NewCoreAPIs("local", "1.1.1", "build", &storage, nil, nil, nil)
+	coreAPIs := core.NewCoreAPIs("local", "1.1.1", "build", &storage, nil, nil, nil, "", "", nil)
 
 	l := logs.NewLogger("test", nil).NewLog("1", logs.RequestContext{})
 	got := coreAPIs.Services.SerGetCommonTest(l)
@@ -65,7 +65,7 @@ func TestSerGetCommonTest(t *testing.T) {
 
 func TestAdmGetTest(t *testing.T) {
 	storage := genmocks.Storage{}
-	coreAPIs := core.NewCoreAPIs("local", "1.1.1", "build", &storage, nil, nil, nil)
+	coreAPIs := core.NewCoreAPIs("local", "1.1.1", "build", &storage, nil, nil, nil, "", "", nil)
 
 	got := coreAPIs.Administration.AdmGetTest()
 	want := "Admin - test"
@@ -84,7 +84,7 @@ func TestSysCreateGlobalConfig(t *testing.T) {
 	storage.On("GetGlobalConfig").Return(nil, nil)
 	storage.On("CreateGlobalConfig", nil, &model.GlobalConfig{Setting: "setting"}).Return(nil)
 
-	app := core.NewCoreAPIs("local", "1.1.1", "build", &storage, nil, nil, nil)
+	app := core.NewCoreAPIs("local", "1.1.1", "build", &storage, nil, nil, nil, "", "", nil)
 
 	gc, _ := app.System.SysCreateGlobalConfig("setting")
 	if gc == nil {
@@ -98,7 +98,7 @@ func TestSysCreateGlobalConfig(t *testing.T) {
 	storage2.On("GetGlobalConfig").Return(nil, nil)
 	storage2.On("CreateGlobalConfig", nil, &model.GlobalConfig{Setting: "setting"}).Return(errors.New("error occured"))
 
-	app = core.NewCoreAPIs("local", "1.1.1", "build", &storage2, nil, nil, nil)
+	app = core.NewCoreAPIs("local", "1.1.1", "build", &storage2, nil, nil, nil, "", "", nil)
 
 	_, err := app.System.SysCreateGlobalConfig("setting")
 	if err == nil {
@@ -112,7 +112,7 @@ func TestSysCreateGlobalConfig(t *testing.T) {
 func TestSysGetOrganization(t *testing.T) {
 	storage := genmocks.Storage{}
 	storage.On("FindOrganization", "_id").Return(&model.Organization{ID: "_id"}, nil)
-	app := core.NewCoreAPIs("local", "1.1.1", "build", &storage, nil, nil, nil)
+	app := core.NewCoreAPIs("local", "1.1.1", "build", &storage, nil, nil, nil, "", "", nil)
 
 	getOrganization, _ := app.System.SysGetOrganization("_id")
 
@@ -122,7 +122,7 @@ func TestSysGetOrganization(t *testing.T) {
 	// second case error
 	storage2 := genmocks.Storage{}
 	storage2.On("FindOrganization").Return(&model.Organization{ID: "_id"}, nil)
-	app = core.NewCoreAPIs("local", "1.1.1", "build", &storage, nil, nil, nil)
+	app = core.NewCoreAPIs("local", "1.1.1", "build", &storage, nil, nil, nil, "", "", nil)
 
 	err, _ := app.System.SysGetOrganization("_id")
 
@@ -136,7 +136,7 @@ func TestSysGetOrganization(t *testing.T) {
 func TestSysGetOrganizations(t *testing.T) {
 	storage := genmocks.Storage{}
 	storage.On("FindOrganizations").Return([]model.Organization{}, nil)
-	app := core.NewCoreAPIs("local", "1.1.1", "build", &storage, nil, nil, nil)
+	app := core.NewCoreAPIs("local", "1.1.1", "build", &storage, nil, nil, nil, "", "", nil)
 
 	getOrganization, _ := app.System.SysGetOrganizations()
 
@@ -146,7 +146,7 @@ func TestSysGetOrganizations(t *testing.T) {
 	// second case error
 	storage2 := genmocks.Storage{}
 	storage2.On("FindOrganizations").Return([]model.Organization{}, nil)
-	app = core.NewCoreAPIs("local", "1.1.1", "build", &storage, nil, nil, nil)
+	app = core.NewCoreAPIs("local", "1.1.1", "build", &storage, nil, nil, nil, "", "", nil)
 
 	err, _ := app.System.SysGetOrganizations()
 
@@ -159,7 +159,7 @@ func TestSysGetOrganizations(t *testing.T) {
 func TestSysGetApplication(t *testing.T) {
 	storage := genmocks.Storage{}
 	storage.On("FindApplication", "_id").Return(&model.Application{ID: "_id"}, nil)
-	app := core.NewCoreAPIs("local", "1.1.1", "build", &storage, nil, nil, nil)
+	app := core.NewCoreAPIs("local", "1.1.1", "build", &storage, nil, nil, nil, "", "", nil)
 
 	getApplication, _ := app.System.SysGetApplication("_id")
 
@@ -169,7 +169,7 @@ func TestSysGetApplication(t *testing.T) {
 	// second case error
 	storage2 := genmocks.Storage{}
 	storage2.On("FindApplication").Return(&model.Application{ID: "_id"}, nil)
-	app = core.NewCoreAPIs("local", "1.1.1", "build", &storage, nil, nil, nil)
+	app = core.NewCoreAPIs("local", "1.1.1", "build", &storage, nil, nil, nil, "", "", nil)
 
 	err, _ := app.System.SysGetApplication("_id")
 
@@ -182,7 +182,7 @@ func TestSysGetApplication(t *testing.T) {
 func TestSysGetApplications(t *testing.T) {
 	storage := genmocks.Storage{}
 	storage.On("FindApplications").Return([]model.Application{}, nil)
-	app := core.NewCoreAPIs("local", "1.1.1", "build", &storage, nil, nil, nil)
+	app := core.NewCoreAPIs("local", "1.1.1", "build", &storage, nil, nil, nil, "", "", nil)
 
 	getApplications, _ := app.System.SysGetApplications()
 
@@ -192,7 +192,7 @@ func TestSysGetApplications(t *testing.T) {
 	// second case error
 	storage2 := genmocks.Storage{}
 	storage2.On("FindApplications").Return([]model.Application{}, nil)
-	app = core.NewCoreAPIs("local", "1.1.1", "build", &storage, nil, nil, nil)
+	app = core.NewCoreAPIs("local", "1.1.1", "build", &storage, nil, nil, nil, "", "", nil)
 
 	err, _ := app.System.SysGetApplications()
 
@@ -208,7 +208,7 @@ func TestSysGetApplications(t *testing.T) {
 
 func TestEncGetTest(t *testing.T) {
 	storage := genmocks.Storage{}
-	coreAPIs := core.NewCoreAPIs("local", "1.1.1", "build", &storage, nil, nil, nil)
+	coreAPIs := core.NewCoreAPIs("local", "1.1.1", "build", &storage, nil, nil, nil, "", "", nil)
 
 	got := coreAPIs.Encryption.EncGetTest()
 	want := "Enc - test"
@@ -251,7 +251,7 @@ func TestCreateApplication(t *testing.T) {
 
 func TestBBsGetTest(t *testing.T) {
 	storage := genmocks.Storage{}
-	coreAPIs := core.NewCoreAPIs("local", "1.1.1", "build", &storage, nil, nil, nil)
+	coreAPIs := core.NewCoreAPIs("local", "1.1.1", "build", &storage, nil, nil, nil, "", "", nil)
 
 	got := coreAPIs.BBs.BBsGetTest()
 	want := "BBs - test"