+++ core/app_services.go
@@ -113,42 +113,56 @@ func (app *application) serGetCommonTest(l *logs.Log) string {
 	return "Services - Common - test"
 }
 
-func (app *application) serGetAppConfig(appTypeIdentifier string, orgID *string, versionNumbers model.VersionNumbers, apiKey *string) (*model.ApplicationConfig, error) {
-	//get the app type
-	applicationType, err := app.storage.FindApplicationType(appTypeIdentifier)
-	if err != nil {
-		return nil, errors.WrapErrorAction(logutils.ActionFind, model.TypeApplicationType, logutils.StringArgs(appTypeIdentifier), err)
-	}
-	if applicationType == nil {
-		return nil, errors.ErrorData(logutils.StatusMissing, model.TypeApplicationType, logutils.StringArgs(appTypeIdentifier))
-	}
-
-	appID := applicationType.Application.ID
-
-	if orgID == nil || apiKey != nil {
-		err = app.auth.ValidateAPIKey(appID, *apiKey)
+func (app *application) serGetAppConfig(appTypeIdentifier string, appID *string, orgID *string, versionNumbers model.VersionNumbers, apiKey *string) (*model.ApplicationConfig, error) {
+	var appTypeID string
+	var err error
+	if appID == nil || *appID == "" {
+		//get the app type
+		applicationType, err := app.storage.FindApplicationType(appTypeIdentifier)
 		if err != nil {
-			return nil, errors.WrapErrorData(logutils.StatusInvalid, model.TypeAPIKey, nil, err)
+			return nil, errors.WrapErrorAction(logutils.ActionFind, model.TypeApplicationType, logutils.StringArgs(appTypeIdentifier), err)
+		}
+		if applicationType == nil {
+			return nil, errors.ErrorData(logutils.StatusMissing, model.TypeApplicationType, logutils.StringArgs(appTypeIdentifier))
+		}
+
+		appID = &applicationType.Application.ID
+		appTypeID = applicationType.ID
+	}
+
+	if appID == nil {
+		return nil, errors.WrapErrorData(logutils.StatusMissing, model.TypeApplicationID, nil, err)
+	}
+
+	if orgID == nil {
+		if apiKey != nil {
+			err := app.auth.ValidateAPIKey(*appID, *apiKey)
+			if err != nil {
+				return nil, errors.WrapErrorData(logutils.StatusInvalid, model.TypeAPIKey, nil, err)
+			}
+		} else {
+			return nil, errors.WrapErrorData(logutils.StatusMissing, model.TypeOrganizationID, nil, err)
 		}
 	}
 
 	var appOrgID *string
 	if orgID != nil {
-		appOrg, err := app.storage.FindApplicationOrganization(appID, *orgID)
+		appOrg, err := app.storage.FindApplicationOrganization(*appID, *orgID)
 		if err != nil {
 			return nil, errors.WrapErrorAction(logutils.ActionFind, model.TypeApplicationOrganization, &logutils.FieldArgs{"app_id": appID, "org_id": *orgID}, err)
 		}
 		appOrgID = &appOrg.ID
 	}
 
-	appConfigs, err := app.storage.FindAppConfigByVersion(applicationType.ID, appOrgID, versionNumbers)
+	// will return the patchAppConfig with greatest verion less than or equal to the versionNumbers provided
+	_, patchAppConfigs, err := app.storage.FindAppConfigByVersion(*appID, appTypeID, appOrgID, versionNumbers)
 	if err != nil {
 		return nil, errors.WrapErrorAction(logutils.ActionFind, model.TypeApplicationConfig, nil, err)
 	}
 
-	if appConfigs == nil {
-		return nil, errors.WrapErrorData(logutils.StatusMissing, model.TypeApplicationConfig, nil, err)
+	if len(patchAppConfigs) > 0 {
+		return &patchAppConfigs[0], nil
 	}
 
-	return appConfigs, nil
+	return nil, nil
 }