+++ driver/web/apis_default.go
@@ -16,10 +16,13 @@ package web
 
 import (
 	"core-building-block/core"
+	"core-building-block/core/model"
 	Def "core-building-block/driver/web/docs/gen"
 	"encoding/json"
 	"net/http"
+	"strings"
 
+	"github.com/google/go-github/v44/github"
 	"github.com/rokwire/core-auth-library-go/v2/tokenauth"
 	"github.com/rokwire/logging-library-go/logs"
 	"github.com/rokwire/logging-library-go/logutils"
@@ -27,7 +30,9 @@ import (
 
 //DefaultApisHandler handles default APIs implementation - version etc
 type DefaultApisHandler struct {
-	coreAPIs *core.APIs
+	coreAPIs                  *core.APIs
+	githubWebhookRequestToken string
+	githubAppConfigBranch     string
 }
 
 //getVersion gives the service version
@@ -49,7 +54,40 @@ func (h DefaultApisHandler) getOpenIDConfiguration(l *logs.Log, r *http.Request,
 	return l.HttpResponseSuccessJSON(data)
 }
 
+func (h DefaultApisHandler) handleWebhookConfigChange(l *logs.Log, r *http.Request, claims *tokenauth.Claims) logs.HttpResponse {
+	// verify webhook request signature
+	data, err := github.ValidatePayload(r, []byte(h.githubWebhookRequestToken))
+	if err != nil {
+		return l.HttpResponseErrorData(logutils.MessageDataStatus(logutils.ActionValidate), model.TypeWebhookSecretToken, nil, nil, http.StatusBadRequest, false)
+	}
+
+	var requestData model.WebhookRequest
+	err = json.Unmarshal(data, &requestData)
+	if err != nil {
+		return l.HttpResponseErrorAction(logutils.ActionUnmarshal, model.TypeWebhookRequest, nil, err, http.StatusBadRequest, true)
+	}
+	gitRefs := strings.Split(requestData.Ref, "/")
+	if len(gitRefs) == 0 {
+		return l.HttpResponseErrorData(logutils.StatusInvalid, logutils.MessageDataType("github webhook repository branch"), logutils.StringArgs(requestData.Ref), nil, http.StatusBadRequest, false)
+	}
+	branchName := gitRefs[len(gitRefs)-1]
+	if branchName != h.githubAppConfigBranch {
+		return l.HttpResponseErrorData(logutils.StatusInvalid, logutils.MessageDataType("github webhook repository branch"), logutils.StringArgs(branchName), nil, http.StatusBadRequest, false)
+	}
+
+	commits := requestData.Commits
+	if len(commits) < 1 {
+		return l.HttpResponseErrorData(logutils.StatusInvalid, model.TypeGithubCommit, nil, nil, http.StatusBadRequest, false)
+	}
+	err = h.coreAPIs.Default.ProcessGitHubAppConfigWebhook(commits, l)
+	if err != nil {
+		return l.HttpResponseErrorAction(logutils.ActionUpdate, model.TypeAppConfigFromWebhook, nil, err, http.StatusInternalServerError, true)
+	}
+
+	return l.HttpResponseSuccess()
+}
+
 //NewDefaultApisHandler creates new rest services Handler instance
-func NewDefaultApisHandler(coreAPIs *core.APIs) DefaultApisHandler {
-	return DefaultApisHandler{coreAPIs: coreAPIs}
+func NewDefaultApisHandler(coreAPIs *core.APIs, githubWebhookRequestToken string, githubAppConfigBranch string) DefaultApisHandler {
+	return DefaultApisHandler{coreAPIs: coreAPIs, githubWebhookRequestToken: githubWebhookRequestToken, githubAppConfigBranch: githubAppConfigBranch}
 }