+++ main.go
@@ -18,6 +18,7 @@ import (
 	"core-building-block/core"
 	"core-building-block/core/auth"
 	"core-building-block/driven/emailer"
+	"core-building-block/driven/github"
 	"core-building-block/driven/profilebb"
 	"core-building-block/driven/storage"
 	"core-building-block/driver/web"
@@ -72,6 +73,20 @@ func main() {
 	mongoDBAuth := envLoader.GetAndLogEnvVar("ROKWIRE_CORE_MONGO_AUTH", true, true)
 	mongoDBName := envLoader.GetAndLogEnvVar("ROKWIRE_CORE_MONGO_DATABASE", true, false)
 	mongoTimeout := envLoader.GetAndLogEnvVar("ROKWIRE_CORE_MONGO_TIMEOUT", false, false)
+	// webhook configs
+	githubToken := envLoader.GetAndLogEnvVar("GITHUB_TOKEN", false, false)
+	githubWebhookRequestToken := envLoader.GetAndLogEnvVar("GITHUB_APP_CONFIG_WEBHOOK_REQUEST_TOKEN", false, false)
+	githubOrgnizationName := envLoader.GetAndLogEnvVar("GITHUB_APP_CONFIG_ORG_NAME", false, false)
+	githubRepoName := envLoader.GetAndLogEnvVar("GITHUB_APP_CONFIG_REPO_NAME", false, false)
+	githubWebhookConfigPath := envLoader.GetAndLogEnvVar("GITHUB_APP_CONFIG_WEBHOOK_CONFIG_PATH", false, false)
+	githubAppConfigBranch := envLoader.GetAndLogEnvVar("GITHUB_APP_CONFIG_BRANCH", false, false)
+
+	githubAdapter := github.NewGitHubAdapter(githubToken, githubOrgnizationName, githubRepoName, githubWebhookConfigPath, githubAppConfigBranch, logger)
+	err = githubAdapter.Start()
+	if err != nil {
+		logger.Fatalf("Cannot start the GitHub adapter: %v", err)
+	}
+
 	storageAdapter := storage.NewStorageAdapter(mongoDBAuth, mongoDBName, mongoTimeout, logger)
 	err = storageAdapter.Start()
 	if err != nil {
@@ -156,10 +171,10 @@ func main() {
 	}
 
 	//core
-	coreAPIs := core.NewCoreAPIs(env, Version, Build, storageAdapter, auth, systemInitSettings, logger)
+	coreAPIs := core.NewCoreAPIs(env, Version, Build, storageAdapter, githubAdapter, auth, systemInitSettings, githubWebhookRequestToken, githubAppConfigBranch, logger)
 	coreAPIs.Start()
 
 	//web adapter
-	webAdapter := web.NewWebAdapter(env, serviceID, auth.ServiceRegManager, port, coreAPIs, host, logger)
+	webAdapter := web.NewWebAdapter(env, serviceID, auth.ServiceRegManager, port, coreAPIs, host, githubWebhookRequestToken, githubAppConfigBranch, logger)
 	webAdapter.Start()
 }